Fixing a type mismatch between declaration and definition.
git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@20965 b38c6c8f-bee8-465d-8448-9abe6915d496

+++ tracking/trackFindingCDC/fitting/src/CDCAxialStereoFusion.cc
@@ -129,12 +129,12 @@ namespace {
 }
 
 
-double CDCAxialStereoFusion::average(const TVectorD& startParameters,
-                                     const TMatrixDSym& startCovMatrix,
-                                     const TVectorD& endParameters,
-                                     const TMatrixDSym& endCovMatrix,
-                                     TVectorD& avgParameters,
-                                     TMatrixDSym& avgCovMatrix)
+FloatType CDCAxialStereoFusion::average(const TVectorD& startParameters,
+                                        const TMatrixDSym& startCovMatrix,
+                                        const TVectorD& endParameters,
+                                        const TMatrixDSym& endCovMatrix,
+                                        TVectorD& avgParameters,
+                                        TMatrixDSym& avgCovMatrix)
 {
   TMatrixDSym startInvCovMatrix = startCovMatrix;
   startInvCovMatrix.Invert();
@@ -153,14 +153,14 @@ double CDCAxialStereoFusion::average(const TVectorD& startParameters,
 }
 
 
-double CDCAxialStereoFusion::average(const TVectorD& startParameters,
-                                     const TMatrixDSym& startCovMatrix,
-                                     const TMatrixD& startAmbiguityMatrix,
-                                     const TVectorD& endParameters,
-                                     const TMatrixDSym& endCovMatrix,
-                                     const TMatrixD& endAmbiguityMatrix,
-                                     TVectorD& avgParameters,
-                                     TMatrixDSym& avgCovMatrix)
+FloatType CDCAxialStereoFusion::average(const TVectorD& startParameters,
+                                        const TMatrixDSym& startCovMatrix,
+                                        const TMatrixD& startAmbiguityMatrix,
+                                        const TVectorD& endParameters,
+                                        const TMatrixDSym& endCovMatrix,
+                                        const TMatrixD& endAmbiguityMatrix,
+                                        TVectorD& avgParameters,
+                                        TMatrixDSym& avgCovMatrix)
 {
   TMatrixD startAmbiguityMatrixTransposed = startAmbiguityMatrix;
   startAmbiguityMatrixTransposed.T();
